+++ MlbRealTimeStatsHttpTrigger/function.json
