+++ MlbRealTimeStatsHttpTrigger/sample.dat
@@ -1,0 +1,4 @@
+{
+    "team": "DET",
+    "date": "20230615"
+}
