--- tests/functions/detroit-tigers-starting-lineup-http-trigger.test.js
@@ -1,12 +1,0 @@
-import HttpTriggerExample from '../../DetroitTigersStartingLineupHttpTrigger/index.js';
-
-const context = {
-  log: jest.fn(),
-};
-context.log.error = jest.fn();
-
-describe('execute', () => {
-  it('is a function', () => {
-    expect(typeof HttpTriggerExample).toBe('function');
-  });
-});
